// src/app/api/apiClient/Client.ts
export interface BlogImageDto {
    id?: string;
    imageUrl?: string;
    blogPostId?: string | undefined;
}

export interface BlogPostDto {
    id?: string | undefined;
    author: string;
    date: Date;
    content: string;
    title?: string;
    images?: BlogImage[] | undefined;
    category?: string | undefined;
}

export interface BaseEntity {
    createdTimestamp?: Date;
    updatedTimestamp?: Date;
}

export interface BlogImage extends BaseEntity {
    id?: string;
    imageUrl?: string;
    blogPostId?: string | undefined;
}
// ...
export class ApiException extends Error {
    message: string;
    status: number;
    response: string;
    headers: { [key: string]: any; };
    result: any;

    constructor(message: string, status: number, response: string, headers: { [key: string]: any; }, result: any) {
        super();

        this.message = message;
        this.status = status;
        this.response = response;
        this.headers = headers;
        this.result = result;
    }

    protected isApiException = true;

    static isApiException(obj: any): obj is ApiException {
        return obj.isApiException === true;
    }
}